% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{function1}
\alias{function1}
\title{function1 Does Stuff to help in server.R}
\usage{
function1(one, two, three = FALSE)
}
\arguments{
\item{one}{A parameter that does one things}

\item{two, }{a parameter that does two things}

\item{three, }{a parameter that does three things}
}
\value{
Returns a stuff from function
}
\description{
function1 Does Stuff to help in server.R
}
